+++ packages/core/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @atlaskit/pragmatic-drag-and-drop
 
+## 2.0.2
+
+### Patch Changes
+
+- [`2b6d25f854593`](https://bitbucket.org/atlassian/atlassian-frontend-monorepo/commits/2b6d25f854593) -
+  Remove legacy ReactDOM.render/hydrate/unmountComponentAtNode usage from non-production code
+  (tests, demo entries, VR fixtures) as part of the React 19 migration
+
 ## 2.0.1
 
 ### Patch Changes